Support for indexes
Added support for e.g. the Nikkei 225 index, which does not have the
volume and adjusted close columns. I think the code of these kind of
indexes end with ".O", so that is what I am basing this on for now.

Also fixed a bug that would attempt to recalculate the matrix when there
is no matrix yet, i.e. download data from the day of the stock split.